fixed real earnings value

+++ functions/getRealEarnings.py
@@ -12,11 +12,11 @@ def getRealEarnings():
     total_time_delta = 0
     no_stats = 0
     for payout in payouts_list:
-        if payout["time_delta"] == 0:
+        if int(payout["time_delta"]) == 0:
             no_stats += 1
             continue
         total_earnings += float(payout["amount_"])
-        total_time_delta += float(payout["time_delta"])
+        total_time_delta += int(payout["time_delta"])
     avg_earnings = total_earnings/(len(payouts_list)-no_stats)
     avg_time_delta = total_time_delta/(len(payouts_list)-no_stats)
     avg_daily_earnings = avg_earnings/avg_time_delta*24*60*60
